Convert `atan` tests to used interval framework (#1504)
Also includes a fix for f64, which was pushing inputs through a
Float32Array. This has been corrected to use a Float64Array.

Issue #792

+++ src/webgpu/util/conversion.ts
@@ -564,7 +564,7 @@ export class Scalar {
 
 /** Create an f64 from a numeric value, a JS `number`. */
 export function f64(value: number): Scalar {
-  const arr = new Float32Array([value]);
+  const arr = new Float64Array([value]);
   return new Scalar(TypeF64, arr[0], arr);
 }
 /** Create an f32 from a numeric value, a JS `number`. */
